Include groupId only if an artifactId is a common suffix
This is a small follow-up to #68 that only includes the groupId
in the name if an artifactId is a common suffix. If no artifactId
is a common suffix, their names should be unambiguous enough.

I've also noted that `Update.name` is used for branch names, so we're
now using NameResolver only for commit messages and PR titles.