Ignore certain dependencies in the updates cache
Certain dependencies can't be checked for updates in a
`ArtificialProject` because they require settings in the build.

For example, com.ccadllc.cedi:build in `ArtificialProject` leads to the
following error:

```
References to undefined settings:

  *:githubProject from *:pomExtra
((com.ccadllc.cedi.build.BuildSettings) BuildSettings.scala:147)

  *:githubProject from *:githubHttpUrl
((com.ccadllc.cedi.build.BuildSettings) BuildSettings.scala:70)
```

So this ignores certain dependencies that cannot be checked for
an update when building up the updates cache.