Extend multi-step update strategy to all source files
This extends the multi-step update strategy outlined in #236 to all
source files. That means we first try an update strategy on all source
files and stop if it changed any file. Only if it didn't change any file,
we'll try the next strategy and so on.

Also included is a relaxed update strategy that splits artifactIds by
hyphens and uses each part as a search term for updating the version
number. This helps in cases where there is an update for a single update
(e.g. for circe-parser) but the version is defined more generically
(e.g. `val circeVersion = "0.11.1"`).

The code is not pretty but will be improved once this idea proves to be
useful.