Always use Config.gitHubLogin in PR body
This uses Config.gitHubLogin as contact in the PR body if
Config.doNotFork is true. Prior to this change, the login of the repo
in which the PR was created was used in the body of PRs. But the login
in the PR body should be the account responsible for running the
scala-steward instance.