Include repo in head param when searching for PRs
This fixes a bug when searching for already existing PRs in a repo
where scala-steward's repo has a different name because the parent
was moved or renamed after scala-steward forked it.

When the repo names differ, the name must be included into the head
parameter like this:
```
head=scala-steward/scalaz-zio:update/scalaz-core-7.3.0-M30
```
Prior to this change, the parameter looked like this:
```
head=scala-steward:update/scalaz-core-7.3.0-M30
```
But this only works when the names are equal.

One consequence of this bug are duplicated PRs like these:
* https://github.com/zio/zio/pull/991
* https://github.com/zio/zio/pull/983