Ignore dependencies introduced by plugins
This ignores dependencies of a repo which are not defined in the build
files but are introduced by plugins. This allows scala-steward to prune
more repositories it doesn't need to look at. Without this scala-steward
tried to update dependencies which were never defined in the build files
and therefore couldn't be updated.