Add Change datatype (#718)
* Add Change datatype

`Change` indicates if some input value has been changed or not. I think
it is clearer than using `Option` in `replaceSomeInOpt` where `Some`
meant that the input was changed. It is also used in the implementation
of `replaceSomeInAllowedParts` where we're making use of its monoid
instance.

The ultimate goal here was to give `replaceSomeInChange` and
`replaceSomeInAllowedParts` the same signature so that one is a drop-in
replacement for the other.

* Fix style issue: unnamed parameter

* Document replaceSomeInAllowedParts

* Document Change