Ignore changes for git submodules when looking for changes
If the sbt project depends on the other sbt project that is
downloaded using git submodule like this:

```
[submodule "submodule"]
  path = submodule
  url = git@github.com:test/submodule.git
```

(Assuming that `test/submodule` is also a sbt project).

```
lazy val submodule = (project in file("submodule"))
```

Scala Steward will crash when it runs on such repository with
"Changes not staged for commit".

This is because
- scala-steward edits the dependencies in submodule project
- Try to commit it
- But it is impossible to commit the changes in the submodule work tree

This PR enalbe scala-steward to run even in such repositories by
ignoring the diffs in git submodule.
https://git-scm.com/docs/git-status