Replace LoggerAlg with LoggerOps
This is a half-reversal of
https://github.com/fthomas/scala-steward/pull/325 which replaced
LoggerOps with LogAlg. The intention of that PR was reduce the usage of
Sync by introducing DateTimeAlg. But the introduction of LogAlg just
made things more complicated since there were now two values used for
logging, logger and logAlg. LogAlg also unnecessarily combined unrelated
methods with different dependencies into one class with the superset of
all dependencies. So this PR introduces LoggerOps again but uses
DateTimeAlg instead of Sync for the implementation of infoTimed and
infoTotalTime.