Automatically exclude dependencies from update check
If `pruneRepos` is `true`, Scala Steward will check all dependencies of
all repos for newer versions to decide on which repos it needs to be
working on. If a dependency can't be checked for updates with
`SbtAlg.getUpdatesForProject`, the whole update check is slowed down
considerably because of the way `UpdateAlg.checkForUpdates` is currently
implemented. Before this change, there was a static list of dependencies
that should be excluded from the update check to speed it up. With this
change this list is now dynamic and it excludes dependencies that
can't be checked for updates for 21 days from the update check. After
that period they are checked again for updates and if that fails again,
they'll be exluded again for the next 21 days.