Do not refresh caches of failing repos
If `--prune-repos=true`, Scala Steward checks for every repository if
their `RepoCache` is still fresh and refreshes it if necessary. A cache
refresh calls sbt to get a list of all dependencies and is therefore a
costly operation. There are some repositories for which the refresh
fails because of a broken build. These broken repos slow down Steward
because it tries to refresh them on every run. With this change, the
caches of those repositores are not refreshed for 7 days. Note that the
caches are not invalidated if there is an error during the refresh.