Only consider versions with pre-release identifiers as pre-releases
This change allows updates from e.g. 8.0.192-R14 to 12.0.2-R18 (where
the 'R' stands for Release). It is currently not allowed because the
hyphen in the version number makes it a pre-release version and we
don't allow updates to a pre-release of a different version series.
With this change we require that the version acually contains a
pre-release identifier like RC, SNAPSHOT, or alpha to be considered a
pre-release version.