Resolve POMs of sbt plugins
This adds `sbtVersion` and `scalaVersion` attributes to
`coursier.Dependency` to correctly resolve POMs of sbt plugins. With
this change, Scala Steward shows the SCM URL or homepage in the PR body
of sbt plugins that are on Maven Central.

Here is a PR of sbt-sonatype before this change:
https://github.com/fthomas/refined-sjs-example/pull/57

and here with this change:
https://github.com/fthomas/refined-sjs-example/pull/63