Run self checks at program start
This adds a `SelfCheckAlg` which is used to run self checks at program
start. The intention of these self checks is to check that components
work correctly before they are used later in the program. This should
help users tweaking the environment until all components work as
expected. For now it only includes a check of HttpExistenceClient which
might need additional JVM properties if Scala Steward is operated behind
a proxy server.