Add "release-$version" as possible tag for diffs and release notes
When looking for version diffs and release notes we now also check URLs
that contain tags like `release-$version`.

ScalaTest uses this scheme for their tags. Here is an example PR that
now includes links to the version diff and release notes because of this
change: https://github.com/fthomas/refined-sjs-example/pull/67