New StewardPlugin tasks, attempt 2
This contains the changes to StewardPlugin from #851 without the other
changes to the core module. I'm more comfortable digesting the changes
of #851 in piecemeal.

The advantage of this change is that we get rid of the parser for the
dependencyUpdates task, that we get more newer versions from
sbt-updates, and that we don't need to set dependencyUpdatesFailBuild
any more because dependencyUpdatesData is not using it.