Do not grow line buffer in slurp unbounded
If a program created by `process.slurp` outputs a lot of lines, the
internal line buffer can grow until there is no more available heap
space and an OOME occurs. I've observed this with a Scalafix migration
that printed a lot of stack traces.

To prevent OOMEs we put an upper bound on the number of lines we're
storing by using the new `appendBounded`.