Do not use RepoCacheRepository in SbtAlg
This is another change inspired by #851 that removes
`RepoCacheRepository` from `SbtAlg`. In `getUpdatesForRepo` we also
need the dependencies of the repo to find updates that change the
groupId. Prior to this change, `RepoCacheRepository` was used to get the
dependencies. The disadvantage of this approach was, that it required a
fresh cache but the cache is only populated and fresh if
`--prune-repos=true`.

We now query sbt for dependencies and updates at the same time so there
is no more need for the `RepoCacheRepository` and updates with a new
groupId are now also reported if `--prune-repos=false`.