Set TTL of Coursier's FileCache to 1 hour
The default TTL of Coursier's FileCache is 24 hours. Now that we're
using Coursier to find new versions, 24 hours is too long since the
runtime of my public instance is about 5-6 hours. If the TTL is greater
than that, Scala Steward could prune repos that have updates available.

This PR sets the TTL to 1 hour. This is fine for my public instance and
I assume that nobody runs Scala Steward more frequent than every hour.