Do not use show to display the output of sbt tasks
Using show in combination with "+" issues this warning sometimes:
```
[warn] Issuing a cross building command, but not all sub projects have
the same cross build configuration. This could result in subprojects
cross building against Scala versions that they are not compatible with.
Try issuing cross building command with tasks instead, since sbt will be
able to ensure that cross building is only done using configured project
and Scala version combinations that are configured.
```

In this case, `+ show stewardDependencies` only list dependencies
for `scalaVersion`. Not using show and printing in our tasks directly,
seems to have the desired effect of displaying dependencies for all
`crossScalaVersions`.