Only check dependencies whose version is in source files
For each dependency in a repository we now also cache the files that
contain its current version number. This is then used in PruningAlg
to discard dependencies whose version is not in any file. We don't need
to check for an update of a dependency if we can't bump the version
number later. This change will prune more repositories.

If pruning is the default, we will be able to use this file list later
in EditAlg instead of computing it there again.