Use Coursier to double-check our version ordering
If #1209 is merged, our own `Order[Version]` will become more important.
To prevent that it does something stupid, this adds a double-check which
uses coursier.core.Version to verify that the next version of an update
is greater than the current. Once our own ordering is battle-tested
enough, this can be removed again.