Throttle (concurrent) requests to Maven Central
Scala Steward has become so fast that it is now frequently being
blacklisted by Maven Central despite using Coursier's caching layer.
Previous attempts to prevent that (#1215 and #1216) were't successful
so I'm now trying to limit the rate of calls to Coursier that
potentially result in a call to Maven Central.