Add our own versions cache (#1220)
* Add our own versions cache

The same rationale as for #1218 applies here too. Our own versions cache
allows for a more fine-grained control of which calls to Coursier (and
therefore potentially to Maven Central) are rate limited. Since we are
using the same TTL for our cache and for Coursier's cache, our cache
should be as fresh as Coursier's.

The previous rate limiting could result in limited calls that hit
Coursier's cache and unlimited calls that missed Coursier's cache.
This won't happen with this implementation. So we're faster now and
reduce the chance of being blacklisted by Maven Central again.

* Do not rely on the cached sort order