Do not store versions in VersionsCacheAlg
This renames VersionsCacheAlg to VersionsCacheFacade and changes it to
not store versions anymore. Versions are now always obtained from
Coursier. VersionsCacheFacade's job now is to keep track of when
versions where updated by Coursier and to rate limit calls to Coursier
that hit the network (because of #1218).

The big advantage of this change is that we don't need to build and
maintain our own versions cache (which would always be inferior to
Coursier's cache). For example, until now we've neglected the
repositories where dependencies are available in our cache which would
have lead to incorrect PRs with #1209.

VersionsCacheFacade still ignores resolvers when storing the timestamp
of the last update, but this will only lead to unlimited calls to
Coursier that might hit the network but not to incorrect data and
behaviour.