Cache versions again (#1265)
Versions are now again cached by `VersionsCache` to better control when
a request to a remote repository happens (see https://github.com/fthomas/scala-steward/pull/1259#issuecomment-579591682 for a reason why we want to do this).

This partly reverts the changes of https://github.com/fthomas/scala-steward/pull/1241.
An improvement over the last time we cached the versions ourself is that
we are now taking resolvers into account. Ignoring resolvers was one of
the main motivations for creating #1241 in the first place.