Push in mergeAndApplyAgain even if there is only a merge commit
Prior to this change `mergeAndApplyAgain` only pushed changes if
applying an update again resulted in changes in the working copy
of a repo. This prevented Scala Steward from updating PRs if
`updatePullRequests` was set to `"always"` but there were no
merge conflicts between the base branch and the update branch.
In these cases, merging the base branch is the only thing required
to bring the update branch up-to-date with the base branch.

With this change Scala Steward now pushes changes even if updating
only results in a merge commit.

Closes #1511.