Prevent updates to less stable pre-release versions
This prevents updates in the same series where the current and new
versions both contain pre-release identifiers and the new version
contains a pre-release identifier with a smaller order than the
pre-release identifer of the current version. It is prevented since the
new version is considered less stable than the current version. We are
preventing updates to pre-releases of a different series for the same
reason.

An example for an update that is prevented by this change is from
2.1.4.0-RC17 to 2.1.4.0-RC17+1-307f2f6c-SNAPSHOT.

This is a follow-up to #1549.