Delete file in createTemporarily if writeFile fails
`FileAlg.createTemporarily` could leave partially written files behind
if `writeFile` fails with an error - for example if it fails with
an `IOException` if there is no space left on the device. With this
change `deleteForce` will be called in case of an error and the error
will be rethrown so that `createTemporarily` fails too (as before).