Grab PRs from the repo, keep only obsolete information
This one is odd for me because I'm not sure what's a cross version. A group
update is an update for multiple dependency of the same group. A cross
dependency is set of depedency that belongs to the same group. So I would
expect a Update.Single to be for one dependency and a Update.Group to be
for one crossDependency