Replace @daily, @weekly, @monthly with arbitrary timespans
This replaces `PullRequestFrequency.{Daily, Weekly, Monthly}` with a new
`Timespan` case for arbitrary timespans. It allows for example to set
the PR frequency like this:
```
pullRequests.frequency = "14 days"
```

`@daily`, `@weekly`, `@monthly` can still be parsed and are translated
to `Timespan(1.day)`, `Timespan(7.days)`, and `Timespan(30.days)`. The
main motivation for this change is to make it easier for users to set
the frequencey to every x days without resorting to cron expressions
which are harder to read and write.