Close obsolete PRs only if a PR was created or updated
This closes obsolete PRs only if a new PR was created or if an existing
PR was updated. Sometimes `processUpdate` does not create a new PR
because it can't figure out where to bump the version number. I think
that in these cases Scala Steward should not close the PRs of older
updates.