Do not check in StewardPlugin if dependencies are defined in build files
This reverts #584 which shouldn't be needed any more since #1188.

The goal of #584 was to ignore dependencies that are added by plugins
because those dependencies cannot be updated. Unfortunately this
solution is brittle because it also ignores plugins that are added via
`addSbtPlugin` in `settings`, see https://github.com/http4s/sbt-http4s-org/pull/26#issuecomment-712351259.
(They also have a source position like ` LinePosition(Defaults.scala,...)`.)

A similar optimization was later added in #1188 which ignores all
dependencies whose versions number cannot be found in files of the repo.
This optimization also covers dependencies that are added by plugins.
Let's see if this is good enough to also cover #584.