Allow passing environment variables to execSandboxed
This allows to pass extra environment variables to `execSandboxed` in
preparation for a fix for #1698. It also changes how environment
variables are used if firejail is enabled. All environment variables
(those from the config and those passed to `execSandboxed`) are
now passed via the `--env` parameter of the firejail command instead of
injecting them into the environment. This has been suggested earlier in
https://github.com/scala-steward-org/scala-steward/pull/337#discussion_r264052558 by @mwz.