Do not put sbt in offline mode
Some builds cannot be loaded in offline mode because of missing
dependencies. They then fail with errors like this:
```
sbt.librarymanagement.ResolveException: unresolved dependency: com.github.gseitz#sbt-release;1.0.13: not found
```

Using `offline` was added in #1253 in an attempt to reduce traffic with
Maven Central. But it turned out that the policy of our version cache was
responsible for most of the traffic (#1259) so I'm fine with removing
this again now.