Close standard input of processes created by slurp
This closes the standard input of processes created by `process.slurp`
so that these processes do not wait for user input. This emulates the
effect of sbt's `-batch` option which does the following:
```sh
  -batch|--batch) exec </dev/null && shift ;; #>
```
So `-batch` redirects `/dev/input` to sbt's stdin which effectively
closes it so that sbt never waits for user input. The `-batch` option
was removed in #1705.

This change is required for projects that fail to load and wait for
user input with:
```
[info] Project loading failed: (r)etry, (q)uit, (l)ast, or (i)gnore?
```