Pass sbt properties as parameters
I didn't realized until seeing this commit
https://github.com/xuwei-k/scala-steward/commit/dbcc2a9094f26bfbb2fde5502b06529f670d7313
by @xuwei-k that sbt properties can also be passed as parameters to the
sbt launcher. I've checked that this works with sbt's default launcher
and Coursier's launcher. I prefer this over supplying them in `JAVA_OPTS`
because it does not interfere with user supplied `JAVA_OPTS`.

Follow-up to #1709.