Change default value of --disable-sandbox
This changes the default value of the `--disable-sandbox` option from
`false` to `true`. I think this is a better default since sandboxed
execution is only useful when the operator of Scala Steward does not
trusts the repositories it is working on. This is only true for public
instances and (almost) never true for private instances. And there are
a lot more private than public instances.

Other arguments for this change are that `firejail` does not work inside
Docker and that documentation and setup become simpler.

To enable sandboxed execution, `--disable-sandbox=false` or the newly
added `--enable-sandbox` option can be used.