Use next version when looking for artifact URLs
This uses dependencies with the next version instead of the current
version when querying Coursier for URLs. If a dependency adds URLs to
its `pom.xml` or `ivy.xml` for its next version Scala Steward will be
able to use them immediately when it is creating PRs for this version.
Currently it can only use them on the next version but one since it
uses the current version to find these URLs.