Add PR number to PullRequestOut and persist it in PullRequestRepository
This contains some of the changes that were also in #1631 and is a
follow-up to #1667. This PR stores the pull request number in
`PullRequestOut` and persists it in `PullRequestRepository`. To sidestep
the issue of migrating old `pull_requests.json` files, the PR number is
added as an optional field to `PullRequestData`.

The goal here is to use the persisted PR number instead of
`VCSExtraAlg.extractPRIdFromUrls` in a future release.