Init Git submodules after syncing with upstream repo
Git submodules are now initialized after the latest changes from the
upstream repo are merged into Scala Steward's fork instead of directly
after cloning our fork and before changes are synced. I hope that this
is enough to ensure that our submodules are always in sync with the
submodules of the upstream repos.

If we're lucky, this closes #339.