Allow updates.limit option to be zero
This allows the repo config option `updates.limit` to be zero so that it
can now be zero or positive instead of positive only. A zero
`updates.limit` means that Scala Steward will not create or update PRs.
It can be used to temporarily disable Scala Steward in a repo or if it
is used in the default repo config it would be similar to a hypothetical
dry-run option.