Add caching layer to pull requests store
The `pull_requests` key value store is repeatedly accessed in
`PruningAlg` with the same key. This happens when we're looking for
already existing PRs which means that the same JSON file is repeatedly
read and decoded. This PR adds a caching layer to the `pull_requests`
store that caches the value of last accessed key.