Show all details in the PR body below each other
This puts `oldVersionDetails` directly above the `ignoreFutureUpdates`
details instead of directly under the release releated URLs.

https://github.com/scala-steward-org/test-repo-2/pull/14#issuecomment-751992453
shows how this looks in comparison the current arrangement. IMO it is
nicer if all details sections are below each other. On the other hand,
`oldVersionDetails` is not as easily missed if it is directly under
the URLs.