Check Monoid laws for all repoconfig classes (#1857)
* Check Monoid laws for all repoconfig classes

Repoconfig classes are not only semigroups (because they can be merged
with the global repoconfig) but also monoids because the empty
`RepoConfig` should be the identity element with regards to merging
configs. That means that merging with the empty `RepoConfig` should not
change a repo-specific config.

This change promotes all `Semigroup` instances of config classes to
`Monoid`s and checks the monoid laws for all of them.
`UpdatesConfig.mergePin` and `UpdatesConfig.mergeIgnore` needed small
adjustments so that they satisfy the left and right identity laws.