Create commit for Scalafix migration even if it failed (#1878)
https://github.com/johnynek/bosatsu/pull/648 contains changes from a
Scalafix migration that exited with a non-zero status code in the same
commit as the version bump. Migration changes and the version bump
are supposed to be in separate commits but weren't because we're
currently only creating the Scalafix commit if it exited without an
error. This change always commits after a Scalafix migration regardless
of its exit status.