Load artifact migrations from this repository
This splits `ArtifactMigrations` into a loader and finder class. The
former loads the latest artifact migrations from this repository instead
of using the static list of migrations that is shipped as resource in
the JAR. That means that all Scala Steward instances use the artifact
migrations from this repository as soon as they are added here.

The following changes with this PR:
* The `--artifact-migrations` option can now be used multiple times to
  load extra migrations not only from one but multiple files.
* Arguments to the `--artifact-migrations` cannot only be local files
  but also remote URLs.
* Scala Steward prints how many artifact migrations it loaded at
  start-up.

The same has been done for Scalafix migrations in #1650.

Closes: #1963