Run scalafmt before creating the update commit (#2319)
This runs `scalafmt --mode changed` after bumping the version but before
creating the update commit. With `--mode changed` only changed files are
reformatted. This reformatting is different to the `PostUpdateHook` that
runs `scalafmt` on scalafmt updates because formatting changes are not
recorded in a separate commit. This is not (yet) configurable via
`.scala-steward.conf` but that could be added later if needed.

We're also only running scalafmt if the project depends on it.

Example PR: https://github.com/scala-steward-org/test-repo-2/pull/59

Closes: #1889