Add `commit-count` label to PR bodies
This adds the `commit-count` label to PR bodies. If the PR contains one
commit, the label will be `commit-count:1`. If it contains more than one
commit, the label will be `commit-count:n:$n` where `$n` is the number of
commits.

This allows to search for PRs with multiple but unspecified number of
commits and for PRs with a specified number of commits.

Closes #2383.