Change return type of `PruningAlg#needsAttention`
Instead of `(Boolean, List[Update.Single])` where the first element
designate if the second is empty, we use
`Option[Nel[UpdateState.WithUpdate]]` instead. This is a little bit
nicer since we don't need to interpret the `Boolean` but just traverse
the `Option`.

Having the `UpdateState`s available in `StewardAlg` will probably also
come in handy when dealing with #2355. I imagine that some kind of rate
limiter in `StewardAlg` will solve that issue.