Add new version of artifact-migrations file
This adds a new default file for artifacts migrations that does not
contain the `initialVersion` fields. Since older Scala Steward instances
load the list of artifact migrations from this repository, removing
fields is a backward-incompatible change so that we cannot remove them
in the old default file.

See #2455 and #2472 for context.