chore: Exclude Akka artifacts due to change of license
Akka starting from version 2.7.0 will use the BSL license, which will require from large companies to pay for license. Due to that any library that uses Akka and is included in a project would also need to pay for licenses.

To avoid unexpected costs for companies, let's add Akka 2.7.0 as ignored for Scala Steward, so that users will need to update it manually and consiously accept the new license.

Not sure how to properly allow for the updates if the user wants it, but if it proves to be difficult we can add some additional settings.

For now it should be possible to override the default ignored dependencies if I am not mistaken? Let me know what everyone thinks about this change.