Fail when Mill can't extract dependencies JSON
If the output from the sandboxed Mill process exceed the configured buffer size, we lose the start of the JSON, leading to `extracted.dropWhile(!_.startsWith("{")).mkString("\n")` evaluating to an empty string, which causes other failures down the line. Those errors don't indicate the actual cause. As of this change, we fail early, right when the actual problem occurs.

Joint work with @IdiosApps.