Add sbt plugins locally and not globally
`StewardPlugin.scala` is the sbt plugin that outputs all relevant
information from a build for consumption by Scala Steward. That plugin
has been installed so far globally at the start of Scala Steward into
sbt's home directory. With this PR, that plugin is now installed into
the local `project/` directories as long as it is needed to run `sbt
stewardDependencies`.

This change is motivated by #2847 - i.e. only sbt >= 1.3.0 can currently
load our sbt plugin. With this change we would be able to load different
plugins dependending on a repo's sbt version.