Use sbt plugins from scala-steward-org/sbt-plugin repo
Instead of including the sbt plugins for extracting dependencies in
this repo we now download them from https://github.com/scala-steward-org/sbt-plugin
and add them as resources to the `core` project as before. This allows to
trim the build a little bit and remove 2.12 from `crossScalaVersions`
which halves the number of CI builds.