Optimize regexes in VersionPositionScanner
I played with using `VersionPositionScanner` instead of
`GitAlg#findFilesContaining` in `RepoCacheAlg` and noticed that the
regex used in `findMillDependency` is quite slow since it allows
whitespace (including newlines) in the groupId and artifactId
identifiers.

This change disallows whitespace in those identifiers and requires that
they are not empty. It increases performance significantly even in the
toy example in the added benchmark.