Rename `WorkspaceAlg#cleanWorkspace` (#2881)
* Rename `WorkspaceAlg#cleanWorkspace`

This renames `WorkspaceAlg#cleanWorkspace` to `cleanReposDir` since this
is what this method actually does. The log message is also changed from
"Clean workspace" to "Clean $reposDir". The former message is misleading
since the workspace needs to be persisted between runs for certain
features to work, but Scala Steward says on every run that it is
cleaning this workspace.

* Rename mkRepoDir to toFile