Support artifact migrations in `Update.ForGroupId`
This supports artifact migrations of multiple artifacts in one update.
Prior to this change, this was not possible because the `newerGroupId`
and `newerArtifactId` fields were discarded while `ForArtifactId`
updates were grouped into `ForGroupId`. With this change, `ForGroupId`
is just a non-empty list of `ForArtifactId`s, so that these fields are
not lost.

Closes: #1974