Add `ForgeSpecificCfg` trait
This adds the `Config.ForgeSpecificCfg` trait which allows us to only
pass config options to `ForgeSelection` that are relevant for the
different `ForgeApiAlg`s instead of the complete `Config` object.

This simplifies the tests a little bit because we don't need create
copies of `Config` to change forge specific configuration.