Use `30 days` instead of `@monthly` in repo config example
This changes the value of the `pullRequests.frequency` config in the
example to slow down future updates in PR bodies from `@monthly`
(which is an alias for `30 days`) to `30 days`. An advantage of this
new value is that it is immediately clear how it needs to be adapted
if you want to receive updates every 7, 14, 60, etc. days.

With `@monthly` you need to look at the documentation to learn what
other special values exists and how to configure a custom frequency
of every n days.