Throw if slurp reads more lines than `max-buffer-size` (#2940)
* Throw if slurp reads more lines than `max-buffer-size`

What has been done?

This changes `slurp` such that it throws an `ProcessBufferOverflowException`
if a process outputs more lines than the max buffer size of the internal
lines buffer. Throwing that exception can be prevented via the new
`SlurpOption.IgnoreBufferOverflow`. `slurp` then behaves as before this
change.

Some `exec` calls now use `SlurpOption.IgnoreBufferOverflow` in cases
where we ignore the output of a process or a partial output of a process
is sufficient.

Why?

This allows us to decide in `ProcessAlg#exec` calls if we need the
complete output of a process or if it is okay to only get a fraction of it
(for example, if we completely ignore the output). In some situations it
is important that we receive the complete output and this allows us to
fail early instead of getting hard to debug errors later that are a
consequence of the incomplete output. The message of
`ProcessBufferOverflowException` tells the user what went wrong and how
the max buffer size can be increased.

History and design constraints:

- `ProcessAlg#exec` returns a list of outputted lines, so that other
  parts of Scala Steward can easily work with the output of external
  processes.
- External processes can run indefinitely and output an indefinite
  number of bytes. This is why timeouts (#776) and a bounded lines
  buffer (#1123) have been added to `slurp`.
- The output of processes is valuable for reporting errors, even if
  that output is not used by the rest of the program.

Alternative to: #2717
Closes: #2718

* Add a predefined Set(IgnoreBufferOverflow) with the correct type

* Add SlurpOptions type alias

* Test what happens if line length > buffer size