Ignore file excluded from git
Git has a .gitignore which can be used to exclude files
from getting into the repository. It's pointless to scan
files ignored via this mechanism as we'll fail to stage
any changes to them.

This approach reads one .gitignore are the top of the file
and then use it to check all the paths. This is not _exact_
but I think it's good enough.

I say that this is not exact because Git supports .gitignore
file nested at arbitrary depth in the repository. To implement
that, we'd need to re-implement the `walk` function which is
not worth it in the scope of this issue.