Add GitHub Action Job Summary summarising failing repos
When Scala Steward's GitHub Action is running a job over many repos, and one or
more of those repos fails, Scala Steward will return a failure ExitCode, but
determining _which_ of those repos has failed, and why, requires ponderously
searching the large workflow job logs for strings like 'ERROR Steward
xxx/yyy failed'. This change provides a GitHub Action Job Summary that is
visible in the GitHub UI without digging into the workflow logs:

* https://github.blog/2022-05-09-supercharging-github-actions-with-job-summaries/
* https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#adding-a-job-summary

A Scala Steward run can fail consistently for individual specific repos,
without preventing it from succeeding on the other repos in the same run.
When we see the big red danger cross reported against the workflow, this is
often a concern solely for the team that owns that specific repo - not the
owners of the many other repos that may have been in the run. Failure on a
repo can be caused by a variety of causes, for instance:

* `org.scalasteward.core.forge.github.GitHubException$RepositoryArchived` - if
  the repo has been _archived_, but not removed from the list of repos for
  Scala Steward to process.
* Invalid sbt configuration in the target repo - for instance, obsolete
  resolvers, version conflicts.

These are problems for that repo in general, not specific to Scala Steward -
but still, not of direct interest to the owners of other repos in the
organisation that are actually being successfully processed by Scala Steward.

Although this change is for the benefit of running Scala Steward in
GitHub Actions, the best place for this change to be implemented is in the
main Scala Steward executable (https://github.com/scala-steward-org/scala-steward),
which has the necessary diagnostic information on the Scala Steward run,
rather than in the Scala Steward GitHub Action itself (https://github.com/scala-steward-org/scala-steward-action),
which is focused on _launching_ Scala Steward.