Remove any run-specific files before starting run
To ensure that its workspace is clean, and that unintended files are not
held over from previous runs, Scala Steward already erases the `repos`
folder before starting its run. With new `run-summary.md` file
introduced by https://github.com/scala-steward-org/scala-steward/pull/3071,
we again want to be careful that the summary file from a previous run
isn't picked up in a subsequent run.

https://github.com/scala-steward-org/scala-steward-action/pull/503
addressed this for the particular concern of GitHub Actions, by ensuring
that the file wasn't persisted to cache, but @fthomas pointed out that
this additional point at the start of Scala Steward's code was also an
appropriate place for removing any run-specific files:

https://github.com/scala-steward-org/scala-steward-action/pull/503#issuecomment-1594477950