Fix #3096: Don't ignore dependency-frequency-limits for *grouped* PRs
https://github.com/scala-steward-org/scala-steward/pull/2515 in February 2022
introduced `dependencyOverrides` frequency settings to rate-limit how often
PRs would be raised by Scala Steward for specific dependencies. Unfortunately
this rate-limit isn't currently enforced for _grouped_ PRs (introduced by
https://github.com/scala-steward-org/scala-steward/pull/2714 in October 2002),
so that bug is fixed with this change.

The rate-limiting in `PruningAlg.newPullRequestsAllowed()` was failing because
it was being passed an empty `artifactLastPrCreatedAt` parameter, even when a
previous PR for that artifact definitely _did_ exist. That empty value (rather
than a populated timestamp) was passed to `newPullRequestsAllowed()` in
`PruningAlg.filterUpdateStates()` from the value returned by
`PullRequestRepository.lastPullRequestCreatedAtByArtifact()` - that method was
_only_ finding PRs for _single_ updates, rather than _grouped_ ones, so was
basically saying "Scala Steward's never raised a PR for that artifact before!"
even though it _had_.

The fix here is to ensure that `lastPullRequestCreatedAtByArtifact()` respects
raised PRs that have grouped, as well as single, updates. We do this by changing
the code so that no matter what kind of update we're looking at (eg grouped or
single), we use the new `asSingleUpdates` field on it to give us the individual
artifacts we want to group on, to find the latest timestamp.

A new test case has been added to `PullRequestRepositoryTest` to check that
`PullRequestRepository.lastPullRequestCreatedAtByArtifact()` is behaving as
desired.